Create Root Context in mupdf
They recommend creating a root context for application running
concurrently. This should allow for caching fonts and layers across
runs. I was also dealing with segfaults out of mupdf when creating
images in a threaded server since there were some shared objects the
library creates that aren't safe to access.

Problems with this solution:
- A reliance on `pthread` which isn't supported on all platforms.

- It's creating c locks which may not play nicely with go routes.
  I've been running with this patch applied in my service and haven't
  noticed any problems.

- Might leak memory over time since there isn't currently a way to clean
  the root context.

Was looking to see if this makes sense to clean up and upstream or if
the problems seem to big. Could update the build to link in the
concurrency only on Unix system.

Repo to get this to happen a lot of the time with documents that share
things across pages. Unfortunately I can't share the pdf we found that
did this consistently.

```
package main

import (
	"context"
	"flag"
	"log"
	"os"
	"sync"

	"github.com/gen2brain/go-fitz"
)

func main() {
	ctx := context.Background()
	flag.Parse()
	args := flag.Args()
	if len(args) != 1 {
		log.Fatal(flag.ErrHelp.Error())
	}

	bs, err := os.ReadFile(args[0])
	if err != nil {
		panic(err)
	}

	doc, err := fitz.NewFromMemory(bs)
	if err != nil {
		panic(err)
	}
	defer doc.Close()

	n := doc.NumPage()
	wg := sync.WaitGroup{}
	wg.Add(n)

	for i := 0; i < n; i++ {
		go func(idx int, wg *sync.WaitGroup) {
			defer wg.Done()
			doc, err := fitz.NewFromMemory(bs)
			if err != nil {
				panic(err)
			}
			defer doc.Close()

			_, err = doc.ImageDPI(idx, 300)
			if err != nil {
				panic(err)
			}
			log.Printf("page %d\n", idx+1)
		}(i, &wg)
	}
	wg.Wait()
}
```